casa-6.6.1-17-pipeline-2024.1.0.8 demanded a new numpy 1.24.4 which deprecated np.bool and np.int

+++ admit/util/Spectrum.py
@@ -620,7 +620,7 @@ class Spectrum(object):
         elif isinstance(chans, ma.masked_array):
             self._chans = chans.data.astype(np.int, copy=True)
         elif isinstance(chans, np.ndarray):
-            self._chans = chans.astype(np.int, copy=True)
+            self._chans = chans.astype(np.int32, copy=True)  # deprecating np.int
         else:
             raise
